add authentications and authoriztion

+++ backend/controlers/goalControler.js
@@ -1,15 +1,15 @@
 import asyncHandler from "express-async-handler"
 import goalModel from "../model/goalModel.js"
+import userModel from "../model/userModel.js"
 
 //@desc Get goals
 //@route GET /api/goals
 //@access private
 export const getGoals = asyncHandler(async (req, res) => {
-    const goals = await goalModel.find()
+    const goals = await goalModel.find({user: req.user.id})
 
     res.status(200).json(goals)
 })
-
 
 
 //@desc Set goal
@@ -22,8 +22,8 @@ export const setGoal = asyncHandler(async (req, res) => {
     }
     const goal = await goalModel.create({
         text: req.body.text,
+        user: req.user.id
     })
-
     res.status(200).json(goal)
 })
 
@@ -38,6 +38,20 @@ export const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await userModel.findById(req.user.id)
+
+    //Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await goalModel.findByIdAndUpdate(req.params.id, req.body, {
@@ -59,6 +73,21 @@ export const deleteGoal = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error('Goal not found')
     }
+
+    const user = await userModel.findById(req.user.id)
+
+    //Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await goalModel.findByIdAndDelete(id)
     res.status(200).json({id})
 })
